// readalyse-frontend/src/app/data-access/api/model/pagination.ts
/**
 * Readalyse API
 * Readalyse API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Pagination { 
    page?: number;
    size?: number;
    totalPages?: number;
    totalElements?: number;
}